requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
basicsr==1.3.4.9
beautifulsoup4==4.11.1
bidict==0.22.0
boltons==23.0.0
breeze-connect==1.0.13
bs4==0.0.1
cached-property==1.5.2
cachetools==5.0.0
caffe-ssd==0.1
caffe-ssd-x86==1.0.0.1
certifi==2021.10.8
cfg-load==0.9.0
charset-normalizer==2.0.12
clang==5.0
click==8.1.3
cloudpickle==2.1.0
coda==0.1.0
colorama==0.4.4
contextlib2==21.6.0
cpool==0.2.2
cycler==0.11.0
Cython==0.29.28
dask==2023.3.1
datasets==2.11.0
debtcollector==2.5.0
declxml==0.9.1
decorator==5.1.1
deprecation==2.1.0
dill==0.3.6
distlib==0.3.4
easydict==1.9
einops==0.6.0
et-xmlfile==1.1.0
etils==0.6.0
executing==1.2.0
filelock==3.7.1
flatbuffers==23.5.26
fonttools==4.32.0
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
future==0.18.2
gast==0.4.0
gems==0.4.0
getgit==0.1.0
git-api==1.7.5
git-bash==0.2
git-clone==1.0.6
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.16.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.56.3
grpcio==1.60.0
h5py==3.1.0
huggingface-hub==0.11.1
hydra-core==1.3.2
idna==3.3
image-utils==0.1.6
imageio==2.19.3
imgaug==0.4.0
importlib-metadata==4.11.3
importlib-resources==5.8.0
ipython==8.6.0
iso8601==1.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
keras==2.15.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kornia==0.2.0
libclang==14.0.1
llvmlite==0.40.1
lmdb==1.3.0
locket==1.0.0
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
mpu==0.23.1
multidict==6.0.4
multiprocess==0.70.14
natsort==8.2.0
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.2
nltk==3.7
numba==0.57.1
numpy==1.24.0
oauthlib==3.2.0
omegaconf==2.3.0
onnx==1.12.0
openai-clip==1.0.1
opencv-contrib-python-headless==4.7.0.72
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
oslo.i18n==5.1.0
oslo.utils==4.13.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
partd==1.3.0
pbr==5.9.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.4.0
platformdirs==2.5.2
progressbar==2.5
progressbar2==4.0.0
promise==2.3
prompt-toolkit==3.0.32
protobuf==4.23.4
psutil==5.9.4
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post5
pycocotools==2.0
pyFFTW==0.13.1
Pygments==2.13.0
pyiqa==0.1.7
pymongo==4.1.1
pyparsing==3.0.8
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyreadline3==3.4.1
pysaliency==0.2.21
python-cephlibs==0.94.5.post1
python-dateutil==2.8.2
python-engineio==4.3.2
python-git==2018.2.1
python-socketio==5.6.0
python-utils==3.3.3
pytorch-msssim==0.1.1
pytorch-warmup==0.1.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==5.3.1
qudida==0.0.4
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.18.0
rouge-score==0.1.2
rsa==4.8
schema==0.7.5
scikit-image==0.19.2
scikit-learn==0.22.2.post1
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
shapely==2.0.1
simple-term-menu==1.4.1
simplejson==3.17.6
six==1.15.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
sporco==0.2.1
stack-data==0.6.0
super-image==0.1.7
sympy==1.12
tb-nightly==2.10.0a20220619
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.15.0
tensorflow-datasets==4.6.0
tensorflow-estimator==2.15.0
tensorflow-gpu==2.9.1
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.9.0
tensorlayer==2.2.5
tensorlayerx @ git+https://github.com/tensorlayer/tensorlayerx.git@66d49ba8de55887eead252167601083141913fc8
termcolor==1.1.0
tf-estimator-nightly==2.10.0.dev2022061908
tf-nightly==2.10.0.dev20220628
tf-nightly-gpu==2.10.0.dev20220619
tf-slim==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
tk==0.1.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchcontrib==0.0.2
torchsummary==1.5.1
torchvision==0.15.1+cu117
tqdm==4.64.0
traitlets==5.5.0
transformers==4.25.1
typing_extensions==4.4.0
tzdata==2022.1
tzlocal==4.2
ultralytics==8.0.160
upgrade-pip==0.1.4
urllib3==1.26.15
virtualenv==20.15.1
wcwidth==0.2.5
websocket-client==1.3.3
Werkzeug==2.1.1
wrapt==1.12.1
xlrd==2.0.1
xxhash==3.2.0
yapf==0.32.0
yarl==1.9.2
zipp==3.15.0
